player: use double for a, b positions

+++ include/vlc_player.h
@@ -787,8 +787,8 @@ vlc_player_SetAtoBLoop(vlc_player_t *player, enum vlc_player_abloop abloop);
  * @return A to B loop status
  */
 VLC_API enum vlc_player_abloop
-vlc_player_GetAtoBLoop(vlc_player_t *player, vlc_tick_t *a_time, float *a_pos,
-                       vlc_tick_t *b_time, float *b_pos);
+vlc_player_GetAtoBLoop(vlc_player_t *player, vlc_tick_t *a_time, double *a_pos,
+                       vlc_tick_t *b_time, double *b_pos);
 
 /**
  * Navigate (for DVD/Bluray menus or viewpoint)
